refactor: Product 인터페이스 엔티티로 경로 변경
