+++ src/basic/features/product/actions/findProductById.ts
@@ -1,4 +1,4 @@
-import { Product } from '../types/Product';
+import { Product } from '../../../shared/entity/model/Product';
 
 const findProductById = (id: Product['id'], productList: Product[]) => {
   return productList.find((product) => product.id === id);
