+++ src/basic/features/product/components/ProductOptionsView.ts
@@ -1,4 +1,4 @@
-import { Product } from '../types/Product';
+import { Product } from '../../../shared/entity/model/Product';
 
 const renderProductOptionsView = (
   SelectView: HTMLSelectElement,
