+++ src/basic/shared/entity/model/Product.ts
